rename files to all lowercase
