+++ src/index.js
@@ -1,4 +1,4 @@
-const ReadingList = require('./readingList')
+const ReadingList = require('./reading-list')
 
 readingList = new ReadingList()
 readingList.open()
